backend/requirements.txt
fastapi[standard]
supabase
pydantic
python-dotenv
